requirements.txt
selenium==4.23.1
vaderSentiment==3.3.2
webdriver_manager==4.0.2
python-dotenv==1.0.1
networkx==3.2.1
matplotlib==3.8.4
